+++ proofs/clock.php
@@ -1,7 +1,11 @@
 <?php
 declare(strict_types = 1);
 
-use Innmind\TimeContinuum\Clock;
+use Innmind\TimeContinuum\{
+    Clock,
+    Period,
+    Format,
+};
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 use Innmind\BlackBox\Set;
 
@@ -34,6 +38,11 @@ return static function() {
                 ->greaterThanOrEqual(0)
                 ->lessThanOrEqual(999);
             $assert
+                ->number($point->microsecond()->toInt())
+                ->int()
+                ->greaterThanOrEqual(0)
+                ->lessThanOrEqual(999);
+            $assert
                 ->number($point->minute()->toInt())
                 ->int()
                 ->greaterThanOrEqual(0)
@@ -54,5 +63,58 @@ return static function() {
                 ->greaterThanOrEqual(365)
                 ->lessThanOrEqual(366);
         },
+    );
+
+    yield proof(
+        'Point in times precision is down to the microsecond',
+        given(PointInTime::any()),
+        static function($assert, $point) {
+            $assert->false(
+                $point->equals(
+                    $point->goBack(Period::microsecond(1)),
+                ),
+            );
+            $assert->false(
+                $point->equals(
+                    $point->goForward(Period::microsecond(1)),
+                ),
+            );
+
+            $assert->true(
+                $point->goForward(Period::microsecond(1))->aheadOf(
+                    $point,
+                ),
+            );
+            $assert->true(
+                $point->aheadOf(
+                    $point->goBack(Period::microsecond(1)),
+                ),
+            );
+        },
+    );
+
+    yield proof(
+        'Clock::at() returns nothing for invalid strings',
+        given(
+            Set\Unicode::strings(),
+            Set\Elements::of(
+                null,
+                Format::cookie(),
+                Format::iso8601(),
+                Format::rfc1036(),
+                Format::rfc1123(),
+                Format::rfc2822(),
+                Format::rfc822(),
+                Format::rfc850(),
+                Format::rss(),
+                Format::w3c(),
+            ),
+        ),
+        static fn($assert, $string, $format) => $assert->null(
+            Clock::live()->at($string, $format)->match(
+                static fn($point) => $point,
+                static fn() => null,
+            ),
+        ),
     );
 };