+++ src/ElapsedPeriod.php
@@ -3,93 +3,58 @@ declare(strict_types = 1);
 
 namespace Innmind\TimeContinuum;
 
-use Innmind\Immutable\Maybe;
-
 /**
  * @psalm-immutable
  */
 final class ElapsedPeriod
 {
     /** @var int<0, max> */
-    private int $milliseconds;
+    private int $microseconds;
 
-    private function __construct(int $milliseconds)
+    private function __construct(int $microseconds)
     {
-        if ($milliseconds < 0) {
-            throw new \RuntimeException((string) $milliseconds);
+        if ($microseconds < 0) {
+            throw new \RuntimeException((string) $microseconds);
         }
 
-        $this->milliseconds = $milliseconds;
+        $this->microseconds = $microseconds;
     }
 
     /**
      * @psalm-pure
+     * @internal
      *
      * @throws \RuntimeException
      */
-    public static function of(int $milliseconds): self
+    public static function of(int $microseconds): self
     {
-        return new self($milliseconds);
+        return new self($microseconds);
     }
 
     /**
-     * @psalm-pure
-     *
-     * @param int<0, max> $milliseconds
+     * @return int<0, max>
      */
-    public static function literal(int $milliseconds): self
-    {
-        return new self($milliseconds);
-    }
-
-    /**
-     * @psalm-pure
-     *
-     * @return Maybe<self>
-     */
-    public static function maybe(int $milliseconds): Maybe
-    {
-        try {
-            return Maybe::just(new self($milliseconds));
-        } catch (\RuntimeException $e) {
-            /** @var Maybe<self> */
-            return Maybe::nothing();
-        }
-    }
-
-    /**
-     * @psalm-pure
-     *
-     * @throws \LogicException When using a period containing months or years
-     */
-    public static function ofPeriod(Period $period): self
-    {
-        if ($period->months() !== 0 || $period->years() !== 0) {
-            // a month or a year is not constant
-            throw new \LogicException('Months and years can not be converted to milliseconds');
-        }
-
-        $milliseconds = Period\Value::day->milliseconds($period->days()) +
-            Period\Value::hour->milliseconds($period->hours()) +
-            Period\Value::minute->milliseconds($period->minutes()) +
-            Period\Value::second->milliseconds($period->seconds()) +
-            $period->milliseconds();
-
-        return new self($milliseconds);
-    }
-
     public function milliseconds(): int
     {
-        return $this->milliseconds;
+        /** @var int<0, max> */
+        return (int) ($this->microseconds / 1_000);
+    }
+
+    /**
+     * @return int<0, max>
+     */
+    public function microseconds(): int
+    {
+        return $this->microseconds;
     }
 
     public function longerThan(self $period): bool
     {
-        return $this->milliseconds > $period->milliseconds();
+        return $this->microseconds > $period->microseconds;
     }
 
     public function equals(self $period): bool
     {
-        return $this->milliseconds === $period->milliseconds();
+        return $this->microseconds === $period->microseconds;
     }
 }