+++ src/Move/Month.php
@@ -27,7 +27,8 @@ final class Month
             Period::hour($point->hour()->toInt())
                 ->add(Period::minute($point->minute()->toInt()))
                 ->add(Period::second($point->second()->toInt()))
-                ->add(Period::millisecond($point->millisecond()->toInt())),
+                ->add(Period::millisecond($point->millisecond()->toInt()))
+                ->add(Period::microsecond($point->microsecond()->toInt())),
         );
 
         if ($newPoint->month()->numberOfDays() < $point->day()->ofMonth()) {