+++ src/PointInTime/Microsecond.php
@@ -1,0 +1,45 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\TimeContinuum\PointInTime;
+
+/**
+ * @psalm-immutable
+ */
+final class Microsecond
+{
+    /**
+     * @param int<0, 999> $microsecond
+     */
+    private function __construct(
+        private int $microsecond,
+    ) {
+    }
+
+    /**
+     * @psalm-pure
+     * @internal
+     *
+     * @param int<0, 999> $microsecond
+     */
+    public static function of(int $microsecond): self
+    {
+        return new self($microsecond);
+    }
+
+    /**
+     * @return int<0, 999>
+     */
+    public function toInt(): int
+    {
+        return $this->microsecond;
+    }
+
+    /**
+     * @return non-empty-string
+     */
+    public function toString(): string
+    {
+        return (string) $this->microsecond;
+    }
+}