+++ tests/PointInTimeTest.php
@@ -23,7 +23,7 @@ class PointInTimeTest extends TestCase
 {
     public function testInterface()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
 
         $this->assertInstanceOf(Year::class, $point->year());
         $this->assertInstanceOf(Month::class, $point->month());
@@ -50,7 +50,7 @@ class PointInTimeTest extends TestCase
 
     public function testPreserveMillisecondsWhenNanosecondsInString()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.999678+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.999678+02:00'));
 
         $this->assertSame(999, $point->millisecond()->toInt());
         $this->assertSame(30, $point->second()->toInt());
@@ -58,7 +58,7 @@ class PointInTimeTest extends TestCase
 
     public function testFormat()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
 
         $this->assertSame(
             '08:01:30 05/10/2016',
@@ -68,7 +68,7 @@ class PointInTimeTest extends TestCase
 
     public function testChangeOffset()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
         $point2 = $point->changeOffset(Offset::of(-2, 30));
 
         $this->assertNotSame($point, $point2);
@@ -92,8 +92,8 @@ class PointInTimeTest extends TestCase
 
     public function testElapsedSince()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
-        $point2 = PointInTime::at('2016-10-05T08:03:30.234+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
+        $point2 = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:03:30.234+02:00'));
         $elapsed = $point2->elapsedSince($point);
 
         $this->assertInstanceOf(ElapsedPeriod::class, $elapsed);
@@ -102,8 +102,8 @@ class PointInTimeTest extends TestCase
 
     public function testAheadOf()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
-        $point2 = PointInTime::at('2016-10-05T08:03:30.234+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
+        $point2 = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:03:30.234+02:00'));
 
         $this->assertTrue($point2->aheadOf($point));
         $this->assertFalse($point->aheadOf($point2));
@@ -111,9 +111,9 @@ class PointInTimeTest extends TestCase
 
     public function testEquals()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
-        $point2 = PointInTime::at('2016-10-05T08:03:30.234+02:00');
-        $point3 = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
+        $point2 = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:03:30.234+02:00'));
+        $point3 = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
 
         $this->assertTrue($point->equals($point3));
         $this->assertFalse($point->equals($point2));
@@ -121,9 +121,9 @@ class PointInTimeTest extends TestCase
 
     public function testGoForward()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
         $point2 = $point->goForward(
-            Period::of(1, 1, 1, 1, 1, 30, 878),
+            Period::of(1, 1, 1, 1, 1, 30, 878, 0),
         );
 
         $this->assertSame(2016, $point->year()->toInt());
@@ -144,9 +144,9 @@ class PointInTimeTest extends TestCase
 
     public function testGoBack()
     {
-        $point = PointInTime::at('2016-10-05T08:01:30.123+02:00');
+        $point = PointInTime::at(new \DateTimeImmutable('2016-10-05T08:01:30.123+02:00'));
         $point2 = $point->goBack(
-            Period::of(1, 1, 1, 1, 1, 30, 125),
+            Period::of(1, 1, 1, 1, 1, 30, 125, 0),
         );
 
         $this->assertSame(2016, $point->year()->toInt());
@@ -174,7 +174,7 @@ class PointInTimeTest extends TestCase
 
     public function testGoBackOneDay()
     {
-        $point = PointInTime::at('2018-03-04');
+        $point = PointInTime::at(new \DateTimeImmutable('2018-03-04'));
         $point2 = $point->goBack(Period::day(1));
         $format = Format::of('Y-m-d\TH:i:s.u');
 
@@ -190,7 +190,7 @@ class PointInTimeTest extends TestCase
 
     public function testGoBackOneMillisecondWhenCurrentPointIsAtPreciselyZeroMillisecond()
     {
-        $point = PointInTime::at('1402-07-21 02:42:53.000000');
+        $point = PointInTime::at(new \DateTimeImmutable('1402-07-21 02:42:53.000000'));
         $point2 = $point->goBack(Period::millisecond(1));
         $format = Format::of('Y-m-d\TH:i:s.u');
 
@@ -200,6 +200,22 @@ class PointInTimeTest extends TestCase
         );
         $this->assertSame(
             '1402-07-21T02:42:52.999000',
+            $point2->format($format),
+        );
+    }
+
+    public function testGoBackOneMicrosecondWhenCurrentPointIsAtPreciselyZeroMillisecond()
+    {
+        $point = PointInTime::at(new \DateTimeImmutable('1402-07-21 02:42:53.000000'));
+        $point2 = $point->goBack(Period::microsecond(1));
+        $format = Format::of('Y-m-d\TH:i:s.u');
+
+        $this->assertSame(
+            '1402-07-21T02:42:53.000000',
+            $point->format($format),
+        );
+        $this->assertSame(
+            '1402-07-21T02:42:52.999999',
             $point2->format($format),
         );
     }